requirements.txt
alembic==1.11.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
amqp==5.1.1
APScheduler==3.10.1
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crcmod==1.7
cryptography==41.0.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.17.3
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mistune==2.0.5
oss2==2.18.0
packaging==23.1
Pillow==8.0.1
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.38
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
PyJWT==2.7.0
PyMySQL==1.0.3
pyOpenSSL==23.2.0
pyrsistent==0.19.3
python-alipay-sdk==3.2.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
requests==2.28.0
six==1.16.0
SQLAlchemy==2.0.15
tencentcloud-sdk-python==3.0.905
typing_extensions==4.6.2
tzlocal==5.0.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
